fix(nebular): bare CB19Backend keeps its grid-axis gradients (#2306)
The interpolation coordinate is no longer cast to float32, which was causing
cotangent underflow when scaled by the tiny _lum_scale (~1.25e-46 without
ssp_data). The coordinate now stays in float64 throughout, mirroring the
#1568 data-protection pattern. Not reachable via SEDModel.build (always
supplies ssp_data); the hazard was direct-backend use only.

+++ src/tengri/components/nebular/cloudy_cb19.py
@@ -636,7 +636,12 @@ def _frac_idx(val: float, grid: jnp.ndarray) -> jnp.ndarray:
     idx = jnp.clip(idx, 0, n - 2)
     dx = grid[idx + 1] - grid[idx]
     frac = jnp.where(dx > 0, (val_clipped - grid[idx]) / dx, 0.0)
-    return (idx + frac).astype(jnp.float32)
+    # Keep interpolation coordinate in the grid's native dtype to prevent cotangent
+    # underflow when scaled by tiny _lum_scale (~1.25e-46 without ssp_data). A float32
+    # coordinate scaled by 1.25e-46 has a float32 cotangent that underflows to exactly
+    # 0.0 in the backward pass, silencing all gradients. The grid is float64 by default
+    # in JAX with x64 enabled, matching the #1568 data-protection pattern (#2306).
+    return idx + frac
 
 
 def _interp_7d(
